+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.4
+
+* Cache credentials to file for reuse
+* Generate crypto keys
+
 ## 0.0.3
 
 * Add Netflix Terms of Use to README.md (@johnf)